changed providerId to be a int not long

+++ src/CommitmentsV2/SFA.DAS.CommitmentsV2.Messages/Events/PaymentOrderChangedEvent.cs
@@ -5,7 +5,7 @@ namespace SFA.DAS.CommitmentsV2.Messages.Events
     public class PaymentOrderChangedEvent
     {
         public long AccountId { get; set; }
-        public long[] PaymentOrder { get; set; }
+        public int[] PaymentOrder { get; set; }
 
     }
 }
